refactor: move Enum class to types
