+++ packages/types/src/assignable.ts
@@ -5,3 +5,17 @@ export abstract class Assignable {
         });
     }
 }
+
+export abstract class Enum {
+    enum: string;
+
+    constructor(properties: any) {
+        if (Object.keys(properties).length !== 1) {
+            throw new Error('Enum can only take single value');
+        }
+        Object.keys(properties).map((key: string) => {
+            (this as any)[key] = properties[key];
+            this.enum = key;
+        });
+    }
+}
